// src/interfaces.ts
 export interface ServerResponse<T> {
    readonly Response: T;
    readonly ErrorCode: PlatformErrorCodes;
    readonly ThrottleSeconds: number;
    readonly ErrorStatus: string;
    readonly Message: string;
    readonly MessageData: { [key: string]: string };
  }
// ...
  export interface DestinyManifest {
    readonly version: string;
    readonly mobileAssetContentPath: string;
    readonly mobileGearAssetDataBases: GearAssetDataBaseDefinition[];
    readonly mobileWorldContentPaths: { [key: string]: string };
    /**
     * This points to the generated JSON that contains all the Definitions. Each key is
     * a locale. The value is a path to the aggregated world definitions (warning:
     * large file!)
     */
    readonly jsonWorldContentPaths: { [key: string]: string };
    /**
     * This points to the generated JSON that contains all the Definitions. Each key is
     * a locale. The value is a dictionary, where the key is a definition type by name,
     * and the value is the path to the file for that definition. WARNING: This is
     * unsafe and subject to change - do not depend on data in these files staying
     * around long-term.
     */
    readonly jsonWorldComponentContentPaths: { [key: string]: { [key: string]: string } };
    readonly mobileClanBannerDatabasePath: string;
    readonly mobileGearCDN: { [key: string]: string };
    /**
     * Information about the "Image Pyramid" for Destiny icons. Where possible, we
     * create smaller versions of Destiny icons. These are found as subfolders under
     * the location of the "original/full size" Destiny images, with the same file name
     * and extension as the original image itself. (this lets us avoid sending largely
     * redundant path info with every entity, at the expense of the smaller versions of
     * the image being less discoverable)
     */
    readonly iconImagePyramidInfo: ImagePyramidEntry[];
  }
  
  export interface GearAssetDataBaseDefinition {
    readonly version: number;
    readonly path: string;
  }
  
  export interface ImagePyramidEntry {
    /** The name of the subfolder where these images are located. */
    readonly name: string;
    /** The factor by which the original image size has been reduced. */
    readonly factor: number;
  }